refactor: move other ui-related classes into chatbot.types.ui
